feat: Add CheckPermissions middleware
This commit adds a new middleware called CheckPermissions. This middleware is responsible for checking the permissions of a user before allowing access to certain routes. It checks if the user has the role_id of 1, in which case it allows access without any further checks. Otherwise, it checks if the user has the necessary permissions to access the 'api/bank-accounts' route with the 'POST' method. If the user does not have the necessary permissions, it returns a 401 Unauthorized response.

The CheckPermissions middleware is added to the app/Http/Kernel.php file as a new middleware entry.

The purpose of this change is to enforce permission checks for certain routes and ensure that only authorized users can perform certain actions.